requirements.txt
pdfminer.six==20221105
PyMuPDF==1.23.3
pytesseract==0.3.10
Pillow==10.0.0
scikit-learn==1.3.0
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
nltk==3.8.1
spacy==3.7.2
pandas==2.1.0
scipy==1.11.0
